+++ relayer-cli/src/commands/config/validate.rs
@@ -1,6 +1,7 @@
 use abscissa_core::{Command, Options, Runnable};
 
 use crate::conclude::Output;
+use crate::config;
 use crate::prelude::*;
 
 #[derive(Command, Debug, Options)]
@@ -10,10 +11,11 @@ impl Runnable for ValidateCmd {
     /// Validate the loaded configuration.
     fn run(&self) {
         let config = app_config();
-        info!("Loaded configuration: {:?}", *config);
+        trace!("loaded configuration: {:#?}", *config);
 
-        // TODO: Validate configuration
-
-        Output::with_success().exit();
+        match config::validate_config(&config) {
+            Ok(_) => Output::success("validation passed successfully").exit(),
+            Err(e) => Output::error(format!("{}", e)).exit(),
+        }
     }
 }