+++ relayer/src/config.rs
@@ -2,13 +2,13 @@
 
 pub mod reload;
 
-use std::collections::HashMap;
+use std::collections::{HashMap, HashSet};
 use std::{fmt, fs, fs::File, io::Write, path::Path, time::Duration};
 
 use serde_derive::{Deserialize, Serialize};
 use tendermint_light_client::types::TrustThreshold;
 
-use ibc::ics24_host::identifier::ChainId;
+use ibc::ics24_host::identifier::{ChainId, ChannelId, PortId};
 use ibc::timestamp::ZERO_DURATION;
 
 use crate::error;
@@ -28,6 +28,19 @@ impl GasPrice {
 impl fmt::Display for GasPrice {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{}{}", self.price, self.denom)
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct ChainFilters {
+    pub channels: HashSet<(PortId, ChannelId)>,
+}
+
+impl Default for ChainFilters {
+    fn default() -> Self {
+        Self {
+            channels: HashSet::new(),
+        }
     }
 }
 
@@ -53,7 +66,9 @@ pub mod default {
 }
 
 #[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(deny_unknown_fields)]
 pub struct Config {
+    #[serde(default)]
     pub global: GlobalConfig,
     #[serde(default)]
     pub telemetry: TelemetryConfig,
@@ -98,26 +113,58 @@ impl Default for Strategy {
     }
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct GlobalConfig {
-    #[serde(default)]
-    pub strategy: Strategy,
+/// Log levels are wrappers over [`tracing_core::Level`].
+///
+/// [`tracing_core::Level`]: https://docs.rs/tracing-core/0.1.17/tracing_core/struct.Level.html
+#[derive(Clone, Debug, PartialEq, Deserialize, Serialize)]
+#[serde(rename_all = "lowercase")]
+pub enum LogLevel {
+    Trace,
+    Debug,
+    Info,
+    Warn,
+    Error,
+}
 
-    /// All valid log levels, as defined in tracing:
-    /// https://docs.rs/tracing-core/0.1.17/tracing_core/struct.Level.html
-    pub log_level: String,
+impl Default for LogLevel {
+    fn default() -> Self {
+        Self::Info
+    }
+}
+
+impl fmt::Display for LogLevel {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            LogLevel::Trace => write!(f, "trace"),
+            LogLevel::Debug => write!(f, "debug"),
+            LogLevel::Info => write!(f, "info"),
+            LogLevel::Warn => write!(f, "warn"),
+            LogLevel::Error => write!(f, "error"),
+        }
+    }
+}
+
+#[derive(Clone, Debug, Deserialize, Serialize)]
+#[serde(default, deny_unknown_fields)]
+pub struct GlobalConfig {
+    pub strategy: Strategy,
+    #[serde(default)]
+    pub filter: bool,
+    pub log_level: LogLevel,
 }
 
 impl Default for GlobalConfig {
     fn default() -> Self {
         Self {
             strategy: Strategy::default(),
-            log_level: "info".to_string(),
+            filter: false,
+            log_level: LogLevel::default(),
         }
     }
 }
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
+#[serde(deny_unknown_fields)]
 pub struct TelemetryConfig {
     pub enabled: bool,
     pub host: String,
@@ -135,6 +182,7 @@ impl Default for TelemetryConfig {
 }
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
+#[serde(deny_unknown_fields)]
 pub struct ChainConfig {
     pub id: ChainId,
     pub rpc_addr: tendermint_rpc::Url,
@@ -158,6 +206,8 @@ pub struct ChainConfig {
     #[serde(default)]
     pub trust_threshold: TrustThreshold,
     pub gas_price: GasPrice,
+    #[serde(default)]
+    pub filters: ChainFilters,
 }
 
 /// Attempt to load and parse the TOML config file as a `Config`.