+++ relayer/src/object.rs
@@ -88,6 +88,12 @@ impl Channel {
             self.src_channel_id, self.src_port_id, self.src_chain_id, self.dst_chain_id,
         )
     }
+    pub fn src_port_id(&self) -> &PortId {
+        &self.src_port_id
+    }
+    pub fn src_channel_id(&self) -> &ChannelId {
+        &self.src_channel_id
+    }
 }
 
 /// A packet worker between a source and destination chain, and a specific channel and port.
@@ -112,6 +118,12 @@ impl Packet {
             "packet::{}/{}:{}->{}",
             self.src_channel_id, self.src_port_id, self.src_chain_id, self.dst_chain_id,
         )
+    }
+    pub fn src_port_id(&self) -> &PortId {
+        &self.src_port_id
+    }
+    pub fn src_channel_id(&self) -> &ChannelId {
+        &self.src_channel_id
     }
 }